+++ crates/core/src/flow/mod.rs
@@ -1,2 +1,3 @@
 pub mod install;
 pub mod setup;
+pub mod resolve;
